+++ frontend/src/constants/storageKeys.ts
@@ -1,0 +1,6 @@
+export const LOCAL_KEY_LAST_POSITION = 'CARFFEINE_LAST_POSITION';
+export const LOCAL_KEY_GOOGLE_MAPS_API = 'CARFFEINE_GOOGLE_MAPS_API';
+export const LOCAL_KEY_GOOGLE_MAPS_API_LAST_LOGIN = 'CARFFEINE_GOOGLE_MAPS_API_LAST_LOGIN';
+export const LOCAL_KEY_TOKEN = 'CARFFEINE_TOKEN';
+
+export const SESSION_KEY_REPORTED_STATIONS = 'CARFFEINE_REPORTED_STATIONS';
