+++ frontend/src/hooks/tanstack-query/station-details/reports/useUpdateStationReport.ts
@@ -7,7 +7,8 @@ import { serverStore } from '@stores/serverStore';
 
 import type { Differences } from '@ui/DetailedStationInfo/StationReportConfirmation';
 
-import { DEFAULT_TOKEN, LOCAL_KEY_TOKEN, SERVERS } from '@constants';
+import { DEFAULT_TOKEN, SERVERS } from '@constants';
+import { LOCAL_KEY_TOKEN } from '@constants/storageKeys';
 
 interface fetchReportStationRequest {
   stationId: number;
